Fix file parsing
Fix the assumption that the structures in size.h
were tightly packed. Define packed size macros
for cases where the on-disk size is needed, and
reserve use of sizeof() for cases where an in-
memory copy is being used. Also fix the symbol
comparison logic for DRI/Alcyon symbols in
size.c

Bump the version in size.c

+++ proto.h
@@ -13,18 +13,18 @@ int cvt_int_m2i ( int i );
 void checkpoint ( char *funcname , short ckpoint );
 int dri_symbol_compare ( const void *a , const void *b );
 int coff_symbol_compare ( const void *a , const void *b );
-void read_sec_hdr ( short in_handle , SEC_HDR *section );
-short process_abs_file ( char *fname , short in_handle );
-void read_dri_header ( short in_handle );
-void read_coff_header ( short in_handle );
+void read_sec_hdr ( int in_handle , SEC_HDR *section );
+short process_abs_file ( char *fname , int in_handle );
+void read_dri_header ( int in_handle );
+void read_coff_header ( int in_handle );
 void print_dri_info ( void );
 void print_coff_info ( void );
-void print_dri_symbols ( short fhand );
-void print_coff_symbols( short fhand );
+void print_dri_symbols ( int fhand );
+void print_coff_symbols( int fhand );
 void show_dri_symbol_type( unsigned int symtype );
-int show_bsd_symbol_type( long value, char *str, int symtype, int other, int description );
+int show_bsd_symbol_type( int32_t value, char *str, int symtype, int other, int description );
 void usage(void);
-void main ( short argc , char *argv []);
+void main ( int argc , char *argv []);
 
 /*****************************************************************************
  *	readint.c
